requirements.txt
certifi==2018.8.24
chardet==3.0.4
dj-database-url==0.5.0
Django==2.0.8
django-crispy-forms==1.7.2
django-extensions==2.1.2
django-filter==2.0.0
django-sequences==2.2
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
gunicorn==19.9.0
idna==2.7
Markdown==3.0
Pillow==5.2.0
psycopg2-binary==2.7.5
PyJWT==1.6.4
pytz==2018.5
requests==2.19.1
six==1.11.0
urllib3==1.23
